Processes

Documentation

user_thread :: Bool -> Chan Message -> User -> IO ()

User thread process Takes boolean of debug mode Takes Channel of Message Type Takes singular user responsible for Outputs Nothing

message_thread :: Bool -> Chan Message -> User -> [User] -> IO ()

Message Thread - Runs 10 times per user defined Takes channel to denote when all user messages are completed takes channel to store message created takes user to show who sent the message takes a list of remaining users who can be sent the randomised message

print_messages :: Message -> IO ()

Function used to print out all messages in clean format

count_messages_for_user :: [Message] -> User -> IO ()

Function to print out how many messages were sent to select user Takes list of messages Takes User to see how many messages were received by prints number of messages recieved

full_user_list :: [User]

Full list of users in the system. Returns a pre-populated list of users In a real life situation, this data would be populated through a database

create_user :: Int -> String -> User

Function used to create a new user in simple format

create_message :: UUID -> String -> Int -> User -> User -> Message

Function used to create a new message in simple format

random_val_in_range :: Int -> IO Int

Calculates a random number between 0 and an inputted number

random_message :: Int -> IO String

Generates a random string of characters of given length

gen_uuid :: Int -> UUID

UUID Generator modified from https://stackoverflow.com/questions/58906666/haskell-uuid-generation